+++ CHANGELOG.md
@@ -18,8 +18,8 @@ CHANGELOG
   `--ami-name-prefix`, `--custom-cookbook`, `--post-install`, `--no-public-ip`, `--cluster-template`, `--vpc-id`,
    `--subnet-id`.
 - Add `--image-name`, `--config`, `--region` parameters to `build-image` command.
-- Create S3 bucket per region shared with cluster and image if custom bucket isn't specified instead creating bucket 
-per cluster. 
+- Add `delete-image` command with `--name`, `--region`, `--force` parameters.
+- Add `describe-image` command with `--name`, `--region` parameters.
 - Split head node and compute fleet instance roles and add possibility to configure a different instance role 
   for each queue.
 - Add possibility to configure different security groups for each queue.